Move canAddTemporalAggregation to CompareAggregation utils
