+++ frontend/src/metabase/query_builder/components/CompareAggregations/utils.ts
@@ -90,3 +90,29 @@ export const canSubmit = (
   const areColumnsValid = columns.length > 0;
   return isPeriodValid && areColumnsValid;
 };
+
+export function canAddTemporalCompareAggregation(
+  query: Lib.Query,
+  stageIndex: number,
+  aggregations?: Lib.AggregationClause[],
+): boolean {
+  const aggregationsToUse = aggregations ?? Lib.aggregations(query, stageIndex);
+
+  if (aggregationsToUse.length === 0) {
+    // Hide the "Compare to the past" option if there are no aggregations
+    return false;
+  }
+
+  const breakoutableColumns = Lib.breakoutableColumns(query, stageIndex);
+  const hasAtLeastOneTemporalBreakoutColumn = breakoutableColumns.some(column =>
+    Lib.isTemporal(column),
+  );
+
+  if (!hasAtLeastOneTemporalBreakoutColumn) {
+    // Hide the "Compare to the past" option if there are no
+    // temporal columns to break out on
+    return false;
+  }
+
+  return true;
+}
