+++ frontend/src/metabase/querying/utils/drills/compare-aggregations-drill/compare-aggregations-drill.tsx
@@ -2,7 +2,10 @@ import { t } from "ttag";
 
 import { useDispatch } from "metabase/lib/redux";
 import { setUIControls } from "metabase/query_builder/actions";
-import { CompareAggregations } from "metabase/query_builder/components/CompareAggregations";
+import {
+  canAddTemporalCompareAggregation,
+  CompareAggregations,
+} from "metabase/query_builder/components/CompareAggregations";
 import { trackColumnCompareViaColumnHeader } from "metabase/querying/analytics";
 import type {
   ClickActionPopoverProps,
@@ -13,11 +16,7 @@ import * as Lib from "metabase-lib";
 export const compareAggregationsDrill: Drill<
   Lib.CompareAggregationsDrillThruInfo
 > = ({ drill, question, query, stageIndex, clicked }) => {
-  const aggregations = Lib.aggregations(query, stageIndex);
-  if (
-    !clicked.column ||
-    !canAddTemporalCompareAggregation(query, stageIndex, aggregations)
-  ) {
+  if (!clicked.column || !canAddTemporalCompareAggregation(query, stageIndex)) {
     return [];
   }
 
@@ -71,27 +70,3 @@ export const compareAggregationsDrill: Drill<
     },
   ];
 };
-
-function canAddTemporalCompareAggregation(
-  query: Lib.Query,
-  stageIndex: number,
-  aggregations: Lib.AggregationClause[],
-): boolean {
-  if (aggregations.length === 0) {
-    // Hide the "Compare to the past" option if there are no aggregations
-    return false;
-  }
-
-  const breakoutableColumns = Lib.breakoutableColumns(query, stageIndex);
-  const hasAtLeastOneTemporalBreakoutColumn = breakoutableColumns.some(column =>
-    Lib.isTemporal(column),
-  );
-
-  if (!hasAtLeastOneTemporalBreakoutColumn) {
-    // Hide the "Compare to the past" option if there are no
-    // temporal columns to break out on
-    return false;
-  }
-
-  return true;
-}
