+++ frontend/src/metabase/visualizations/click-actions/actions/CompareAggregationsAction/CompareAggregationsAction.tsx
@@ -3,7 +3,10 @@ import { t } from "ttag";
 import { useDispatch, useSelector } from "metabase/lib/redux";
 import { checkNotNull } from "metabase/lib/types";
 import { setUIControls } from "metabase/query_builder/actions";
-import { CompareAggregations } from "metabase/query_builder/components/CompareAggregations";
+import {
+  canAddTemporalCompareAggregation,
+  CompareAggregations,
+} from "metabase/query_builder/components/CompareAggregations";
 import { getQuestion } from "metabase/query_builder/selectors";
 import { trackColumnCompareViaPlusModal } from "metabase/querying/analytics";
 import type { LegacyDrill } from "metabase/visualizations/types";
@@ -17,14 +20,13 @@ export const CompareAggregationsAction: LegacyDrill = ({
   const query = question.query();
   const stageIndex = -1;
   const { isEditable } = Lib.queryDisplayInfo(query);
-  const aggregations = Lib.aggregations(query, stageIndex);
 
   if (
     !clicked ||
     clicked.value !== undefined ||
     !clicked.columnShortcuts ||
     !isEditable ||
-    !canAddTemporalCompareAggregation(query, stageIndex, aggregations)
+    !canAddTemporalCompareAggregation(query, stageIndex)
   ) {
     return [];
   }
@@ -37,6 +39,7 @@ export const CompareAggregationsAction: LegacyDrill = ({
   }: ClickActionPopoverProps) => {
     const currentQuestion = useSelector(getQuestion);
     const dispatch = useDispatch();
+    const aggregations = Lib.aggregations(query, stageIndex);
 
     function handleSubmit(aggregations: Lib.ExpressionClause[]) {
       const nextQuery = aggregations.reduce(
@@ -83,27 +86,3 @@ export const CompareAggregationsAction: LegacyDrill = ({
     },
   ];
 };
-
-function canAddTemporalCompareAggregation(
-  query: Lib.Query,
-  stageIndex: number,
-  aggregations: Lib.AggregationClause[],
-): boolean {
-  if (aggregations.length === 0) {
-    // Hide the "Compare to the past" option if there are no aggregations
-    return false;
-  }
-
-  const breakoutableColumns = Lib.breakoutableColumns(query, stageIndex);
-  const hasAtLeastOneTemporalBreakoutColumn = breakoutableColumns.some(column =>
-    Lib.isTemporal(column),
-  );
-
-  if (!hasAtLeastOneTemporalBreakoutColumn) {
-    // Hide the "Compare to the past" option if there are no
-    // temporal columns to break out on
-    return false;
-  }
-
-  return true;
-}
